+++ nessi/globopt/ga/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+"""
+nessi.ga
+================================================
+"""
+from __future__ import (absolute_import,
+                        division,
+                        print_function,
+                        unicode_literals)
+
+from .genalg import Genalg
+
+if __name__ == '__main__':
+    import doctest
+    doctest.testmod(exclude_empty=True)