+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - *specfx* method for SUdata()
 - *specfk* method for SUdata()
 - *wiggle* method for SUdata()
+- Standard Genetic Algorithm implementation (mu, pc, pm)
 
 ### Modified
 - SU header structure now fits SU/CWP format (before SEG-Y rev 1)